Fix non unity constraints
