+++ autode/opt/optimisers/__init__.py
@@ -1,15 +1,15 @@
 from autode.opt.optimisers.base import NDOptimiser
 from autode.opt.optimisers.bfgs import BFGSOptimiser
-from autode.opt.optimisers.rfo import RFOOptimiser
-from autode.opt.optimisers.prfo import PRFOOptimiser
-from autode.opt.optimisers.constrained_rfo import CRFOOptimiser
+from autode.opt.optimisers.rfo import RFOptimiser
+from autode.opt.optimisers.prfo import PRFOptimiser
+from autode.opt.optimisers.crfo import CRFOptimiser
 from autode.opt.optimisers.steepest_descent import (CartesianSDOptimiser,
                                                     DIC_SD_Optimiser)
 
 __all__ = ['NDOptimiser',
            'BFGSOptimiser',
-           'RFOOptimiser',
-           'PRFOOptimiser',
-           'CRFOOptimiser',
+           'RFOptimiser',
+           'PRFOptimiser',
+           'CRFOptimiser',
            'CartesianSDOptimiser',
            'DIC_SD_Optimiser']
